requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.0.1
tqdm==4.66.4
playwright==1.46.0
langchain==0.3.13
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
httpx==0.28.1
pinecone-client==5.0.1
